+++ src/net/sourceforge/fenixedu/presentationTier/Action/credits/ShowTeacherCreditsDispatchAction.java
@@ -160,7 +160,7 @@ public class ShowTeacherCreditsDispatchAction extends FenixDispatchAction {
 	double managementCredits = teacher.getManagementFunctionsCredits(executionSemester);
 	double serviceExemptionCredits = teacher.getServiceExemptionCredits(executionSemester);
 	double thesesCredits = teacher.getThesesCredits(executionSemester);
-	int mandatoryLessonHours = teacher.getMandatoryLessonHours(executionSemester);
+	double mandatoryLessonHours = teacher.getMandatoryLessonHours(executionSemester);
 	TeacherService teacherService = teacher.getTeacherServiceByExecutionPeriod(executionSemester);
 	return new CreditLineDTO(executionSemester, teacherService, managementCredits, serviceExemptionCredits,
 		mandatoryLessonHours, teacher, thesesCredits);
